menambahkan fitur export laporan format excel

+++ laporan.php
@@ -2,6 +2,21 @@
 include 'koneksi.php';
 include 'function.php';
 include('template/header.php');
+
+include_once('template/header.php');
+
+if (isset($_POST['tampilkan'])) {
+    $p_awal  = $_POST['p_awal'];
+    $p_akhir = $_POST['p_akhir'];
+
+    $link = "export-laporan.php?cari=true&p_awal=$p_awal&p_akhir=$p_akhir";
+    // query sesuai dengan keyword
+    $buku_tamu = query("SELECT * FROM tamu WHERE tanggal BETWEEN '$p_awal' AND '$p_akhir' ");
+} else {
+    // query ambil semua data buku tamu
+    $buku_tamu = query("SELECT * FROM tamu ORDER BY tanggal DESC");
+}
+
 ?>
 
 <!-- Begin Page Content -->
@@ -64,6 +79,17 @@ include('template/header.php');
   </div>
 
   <div class="card-body">
+    <div class="card shadow mb-4">
+      <div class="card-header py-3">
+        <a href="<?= isset($_POST['tampilkan']) ? $link : 'export-laporan.php'; ?>" target="_blank" class="btn btn-success btn-icon-split">
+          <span class="icon text-white-50">
+            <i class="fas fa-file-excel"></i>
+          </span>
+          <span class="text">Export Laporan</span>
+        </a>
+      </div>
+    </div>
+
     <div class="table-responsive">
       <table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
         <thead>
@@ -85,9 +111,6 @@ include('template/header.php');
             $p_akhir = $_POST['p_akhir'];
             // penomoran auto-increment
             $no = 1;
-
-            // query untuk memanggil semua data dari tabel buku_tamu
-            $buku_tamu = query("SELECT * FROM tamu WHERE tanggal BETWEEN '$p_awal' AND '$p_akhir' ");
             foreach ($buku_tamu as $tamu) : ?>
               <tr>
                 <td><?= $no++; ?></td>
@@ -99,13 +122,11 @@ include('template/header.php');
                 <td><?= $tamu['kepentingan'] ?></td>
                 <td>
                   <a class="btn btn-success" href="edit-tamu.php?id=<?= $tamu['id_tamu'] ?>">Ubah</a>
-                  <a onclick="return confirm('Apakah anda yakin ingin menghapus data ini?')"
-                    class="btn btn-danger"
-                    href="hapus-tamu.php?id=<?= $tamu['id_tamu'] ?>">Hapus</a>
+                  <a onclick="return confirm('Apakah anda yakin ingin menghapus data ini?')" class="btn btn-danger" href="hapus-tamu.php?id=<?= $tamu['id_tamu'] ?>">Hapus</a>
                 </td>
               </tr>
-          <?php endforeach;
-          } ?>
+            <?php endforeach; ?>
+          <?php } ?>
         </tbody>
       </table>
     </div>
@@ -117,4 +138,3 @@ include('template/header.php');
 <?php
 include('template/footer.php');
 ?>
-
